+++ lib/common/widgets/dialogs/job_creator_dialog.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '/common/models/job.dart';
-import '/common/widgets/form_fields/job_form_field.dart';
+import '/common/widgets/form_fields/job_form_field_list_tile.dart';
 import '/misc/form_service.dart';
 
 class JobCreatorDialog extends StatefulWidget {
@@ -32,7 +32,7 @@ class _JobCreatorDialogState extends State<JobCreatorDialog> {
       title: const Text("Ajouter un nouveau métier"),
       content: Form(
         key: _formKey,
-        child: JobFormField(
+        child: JobFormFieldListTile(
           initialValue: Job(),
           onSaved: (Job? job) => setState(() {
             _job = job;
